+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 This CHANGELOG.md file will contain the update log for the latest set of updates to the templates
 
+
+# UPDATES for Master (Release 3.9) - August 6, 2018
+
+1.  Added support for private master nodes
+2.  Addes support for private infra nodes
+3.  Removed inbound NAT rules for master LB to better secure master nodes
+
+
+# UPDATES for Master (Release 3.9) - July 14, 2018
+
+1.  Added support for Accelerated Networking
+2.  Added support for existing or new VNet
+
+
+# UPDATES for Master (Release 3.9) - May 22, 2018
+
+1.  Added parameter for CNS VM Size
+2.  Added support for non-HA masters by allowing a single master
+3.  Cleaned up Azure Cloud Provider configuration
+
+
+# UPDATES for Master (Release 3.9) - May 19, 2018
+
+1.  Updated scripts to support 3.9.27
+2.  Added Support for RHEL 7.5
+3.  Added Container Native Storage (CNS) support
+4.  Added support for custom IP range for the Virtual Network
+
 # UPDATES for Master (Release 3.9) - March 28, 2018
 
 1.  Create Release 3.9 Branch